import/requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# export.py: 12,13
# meta.py: 11,12
SQLAlchemy == 1.3.4

# export.py: 22
# import.py: 12
# meta.py: 13
loguru == 0.3.2

# export.py: 6,21
# import.py: 7
# importData.py: 7
# importData_wStatus.py: 4
# meta.py: 7
numpy == 1.16.4

# export.py: 5
# import.py: 5
# importData.py: 5
# importData_wStatus.py: 3
# meta.py: 6 
pandas == 0.24.2 
              
# export.py: 4
pyodbc == 4.0.26
             
# import.py: 10
# importData_wStatus.py: 5
regex == 2019.06.08
             
# config.py: 1
# export.py: 2
# import.py: 4
# importData.py: 4
# importData_wStatus.py: 2
# meta.py: 4
PyYaml == 5.3

           